lyb FEATURE segment header before subtree metadata
This change provides new possibilities for optimizations. For example,
subtree metadata for non-top-level term nodes does not have to be
written. Or it is also possible to have a optimal LYB format for some
type of node. The change is that the node hash is written as soon as
possible and the subtree metadata optionally follows afterwards.
However, for top-level nodes, the subtree metadata should be written
so that the nodes can be skipped.